+++ src/lib/index.ts
@@ -5,9 +5,9 @@ export { default as CodeEditor } from './components/CodeEditor.svelte'
 export { default as MarkdownIt } from './components/MarkdownIt.svelte'
 export { default as PresentationViewer } from './components/PresentationViewer.svelte'
 export { default as ViewerOverlay } from './components/ViewerOverlay.svelte'
-
 /* attachments */
 export * from './attachments/reactive-styles.js'
+export * from './attachments/slideLock.js'
 
 /* fx  */
 export * from './vfx/typewriter.js'